Implement Codable Protocol and UserDefaults to Save Data
The reasons for this change is to have the application encode the items in the shopping list array and then take that encoded array and store it in NSUserDefaults.  Once I run the application on the simulator and then press shift+command+and the H key twice, I'm able to kill the app and then reopen it from the home screen to see the persisted data.