Implement BaseTableViewController.Swift FIle
In order to subclass viewControllers and use a method called requestInput(), will do all the heavy lifting of creating an alert box and presenting it to get user input.  After subclassing takes place of BaseTableViewController, call the requestInput() method and pass in a title, message, and a handler, which will return the text that was entered as input by the user in the handler.